Update Pigeonhole.cpp
Suppose that a flock of 20 pigeons flies into a set of 19 pigeonholes to roost. Because there are 20 pigeons but only 19 pigeonholes, a least one of these 19 pigeonholes must have at least two pigeons in it. To see why this is true, note that if each pigeonhole had at most one pigeon in it, at most 19 pigeons, one per hole, could be accommodated. This illustrates a general principle called the pigeonhole principle, which states that if there are more pigeons than pigeonholes, then there must be at least one pigeonhole with at least two pigeons in it.
For more info:
https://www.geeksforgeeks.org/discrete-mathematics-the-pigeonhole-principle/
The question of above solution is:
https://codeforces.com/problemset/problem/1536/B 

From Line 13-26 the code is generating lexicographically all possible combination from a to zzzzz.